feat: listページの処理フックスにまとめた
