+++ frontend/nextjs-project/src/app/(app)/list/client.tsx
@@ -1,14 +1,15 @@
 "use client";
 
 import { Box, Typography, Stack } from "@mui/material";
-import React, { useState, useCallback } from "react";
+import React from "react";
 import { useGeolocation } from "@/hooks/useGeolocation";
 import DiscoveryCard from "@/components/ui/DiscoveryCard";
 import { usePosts } from "@/hooks/usePosts";
 import { useFilterStore } from "@/stores/filterStore";
-import { useSearchParams, useRouter, usePathname } from "next/navigation";
+import { useSearchParams } from "next/navigation";
 import LeafyLoader from "@/components/features/loading/LeafyLoader"; // 作成したローダーをインポート
 import { SearchBarOnListPage } from "@/components/features/search/SearchBar";
+import { useListSearchBar } from "@/hooks/useListSearchBar";
 
 import { Post } from "@/types/post";
 
@@ -24,13 +25,9 @@ export default function ListClient({ initialPosts }: ListClientProps) {
     error: geolocationError,
   } = useGeolocation();
   const { sort: currentSort } = useFilterStore();
+  const { searchQuery, handleSearch, handleQueryChange } = useListSearchBar();
   const searchParams = useSearchParams();
-  const router = useRouter();
-  const pathname = usePathname();
   const currentScope = searchParams.get("scope");
-  const [searchQuery, setSearchQuery] = useState<string>(
-    searchParams.get("q") ?? "",
-  );
 
   // TODO: 認証情報を取得
   const user = { uid: "123e4567-e89b-12d3-a456-426614174000" };
@@ -44,29 +41,6 @@ export default function ListClient({ initialPosts }: ListClientProps) {
       query: searchQuery,
     },
     { posts: initialPosts },
-  );
-
-  const handleSearch = useCallback(
-    async (q: string) => {
-      setSearchQuery(q);
-      const params = new URLSearchParams(searchParams.toString());
-      params.set("q", q);
-      router.replace(`${pathname}?${params.toString()}`);
-    },
-    [pathname, router, searchParams],
-  );
-
-  const handleQueryChange = useCallback(
-    (q: string) => {
-      if (q.trim() === "") {
-        setSearchQuery("");
-        const params = new URLSearchParams(searchParams.toString());
-        params.delete("q");
-        const qs = params.toString();
-        router.replace(qs ? `${pathname}?${qs}` : pathname);
-      }
-    },
-    [pathname, router, searchParams],
   );
 
   if (geolocationLoading) {
